+++ django/projects/mysite/__init__.py
@@ -4,4 +4,5 @@ from __future__ import absolute_import, unicode_literals
 # Django starts so that shared_task will use this app.
 from .celery import app as celery_app
 
+# Export celery_app in the mysite module
 __all__ = ['celery_app']