+++ django/projects/mysite/settings_base.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
+from __future__ import absolute_import
 
 # General Django settings for mysite project.
 
@@ -10,6 +11,8 @@ import logging
 import mysite.pipelinefiles as pipelinefiles
 import mysite.utils as utils
 import six
+
+from celery.schedules import crontab
 
 try:
     import psycopg2
@@ -246,15 +249,6 @@ USER_REGISTRATION_ALLOWED = False
 # A new user's defaul groups
 NEW_USER_DEFAULT_GROUPS = []
 
-# A sequence of modules that contain Celery tasks which we want Celery to know
-# about automatically.
-CELERY_IMPORTS = (
-    'catmaid.control.cropping',
-    'catmaid.control.roi',
-    'catmaid.control.treenodeexport',
-    'catmaid.control.nat'
-)
-
 # While pickle can cause security problems [1], we allow it for now and trust
 # that the Celery server will only accept connections from CATMAID. To improve
 # security, this should be changed though, see also [2].
@@ -262,6 +256,20 @@ CELERY_IMPORTS = (
 # [2] https://github.com/catmaid/CATMAID/issues/630
 CELERY_ACCEPT_CONTENT = ['pickle']
 CELERY_TASK_SERIALIZER = 'pickle'
+
+# The default set of periodic tasks
+CELERY_BEAT_SCHEDULE = {
+    # Clean cropped stack directory every night at 23:30.
+    'daily-crop-data-cleanup': {
+        'task': 'catmaid.tasks.cleanup_cropped_stacks',
+        'schedule': crontab(hour=23, minute=30)
+    },
+    # Update project statistics every night at 23:45.
+    'daily-project-stats-summary-update': {
+        'task': 'catmaid.tasks.update_project_statistics',
+        'schedule': crontab(hour=23, minute=45)
+    }
+}
 
 # We use django-pipeline to compress and reference JavaScript and CSS files. To
 # make Pipeline integrate with staticfiles (and therefore collecstatic calls)